moved request helpers to other file
